server/requirements.txt
flask
flask-cors
firebase-admin
groq
google-cloud-firestore
python-dotenv
requests